listener/requirements.txt
requests
psutil
libais
pymongo
psycopg2-binary

